requirements_junk.txt
Django==1.10.5
django-filter==1.0.1
django-imagekit==4.0
django-static-precompiler==1.5
drf-nested-routers==0.11.1
Markdown==2.6.8
model-mommy==1.3.2
Pillow==4.0.0
pytest-django==3.1.2
